refactor: create remote-config module
