+++ src/components/feedback/FeedbackContext.tsx
@@ -1,5 +1,5 @@
 import React, {createContext, useContext, useEffect, useState} from 'react';
-import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
+import {useRemoteConfigContext} from '@atb/modules/remote-config';
 
 export type FeedbackQuestionsViewContext = 'departures' | 'assistant';
