+++ src/mobility/use-vehicles-poll-interval.tsx
@@ -1,4 +1,4 @@
-import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
+import {useRemoteConfigContext} from '@atb/modules/remote-config';
 
 export const useVehiclesPollInterval = () => {
   const {vehicles_poll_interval: pollInterval} = useRemoteConfigContext();
