+++ src/modules/feature-toggles/__tests__/use-feature-toggles-context-state.test.ts
@@ -1,7 +1,7 @@
 import {act, renderHook} from '@testing-library/react-hooks';
 import {useFeatureTogglesContextState} from '../use-feature-toggle-context-state.ts';
 import type {StorageService} from '@atb/modules/storage';
-import type {RemoteConfig} from '@atb/remote-config';
+import type {RemoteConfig} from '@atb/modules/remote-config';
 import {toggleSpecifications} from '../toggle-specifications.ts';
 import {toStorageKey} from '../utils.ts';
 import {FeatureToggleNames} from '../types.ts';
