+++ src/screen-components/loading-screen/use-loading-state.ts
@@ -3,7 +3,7 @@ import {useCallback, useEffect, useRef, useState} from 'react';
 import {LoadingParams, LoadingState} from './types';
 import {useFirestoreConfigurationContext} from '@atb/modules/configuration';
 import {useIsLoadingAppState} from '@atb/screen-components/loading-screen';
-import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
+import {useRemoteConfigContext} from '@atb/modules/remote-config';
 
 export const useLoadingState = (timeoutMs: number): LoadingState => {
   const isLoadingAppState = useIsLoadingAppState();
