+++ src/stacks-hierarchy/Root_LoginAvailableFareContractWarningScreen.tsx
@@ -7,7 +7,7 @@ import {Button} from '@atb/components/button';
 import {ThemeText} from '@atb/components/text';
 import {useFocusOnLoad} from '@atb/utils/use-focus-on-load';
 import {useFareContracts, useTicketingContext} from '@atb/modules/ticketing';
-import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
+import {useRemoteConfigContext} from '@atb/modules/remote-config';
 import {RootStackScreenProps} from '@atb/stacks-hierarchy/navigation-types';
 import {useTimeContext} from '@atb/modules/time';
 import {FullScreenFooter} from '@atb/components/screen-footer';
